+++ src/Electronics.java
@@ -2,7 +2,7 @@ public class Electronics extends Product {
     private String brand;
     private String model;
 
-    public Electronics(String name, float price, int quantity, String brand, String model) {
+    public Electronics(String name, double price, int quantity, String brand, String model) {
         super(name, price, quantity);
         this.brand = brand;
         this.model = model;
